added levels

+++ src/frogger/Turtle.java
@@ -20,7 +20,7 @@ public class Turtle extends MovingObjects{
 	 * @param ypos y-coordinate for the object for its starting point
 	 * @param s speed of the object
 	 * */
-	public Turtle(int xpos, int ypos, int s) {
+	public Turtle(int xpos, int ypos, double s) {
 		super(xpos,ypos, s);
 		setImage(getTurtle2());
 	}
